setting up communication between network and consensus using channels with corresponding direction
